+++ trunk/cpsycle/ui/src/uidef.c
@@ -17,8 +17,8 @@
 
 extern psy_ui_App app;
 
-void psy_ui_setrectangle(psy_ui_Rectangle* self, intptr_t left, intptr_t top, intptr_t width,
-	intptr_t height)
+void psy_ui_setrectangle(psy_ui_Rectangle* self, double left, double top, double width,
+	double height)
 {
    self->left = left;
    self->top = top;
@@ -35,7 +35,7 @@ int psy_ui_rectangle_intersect_rectangle(const psy_ui_Rectangle* self,
 		other->bottom < self->top);
 }
 
-int psy_ui_rectangle_intersect(const psy_ui_Rectangle* self, intptr_t x, intptr_t y)
+int psy_ui_rectangle_intersect(const psy_ui_Rectangle* self, double x, double y)
 {
 	return (x >= self->left && x < self->right && 
 			y >= self->top && y < self->bottom);
@@ -44,13 +44,13 @@ int psy_ui_rectangle_intersect(const psy_ui_Rectangle* self, intptr_t x, intptr_
 // from stackoverflow by metamal
 // todo: use liang-barsky algorithm
 bool psy_ui_rectangle_intersect_segment(const psy_ui_Rectangle* self,
-	intptr_t a_p1x, intptr_t a_p1y, intptr_t a_p2x, intptr_t a_p2y)
+	double a_p1x, double a_p1y, double a_p2x, double a_p2y)
 {
 	// Find min and max X for the segment
-	intptr_t a_rectangleMinX = self->left;
-	intptr_t a_rectangleMinY = self->top;
-	intptr_t a_rectangleMaxX = self->right;
-	intptr_t a_rectangleMaxY = self->bottom;
+	double a_rectangleMinX = self->left;
+	double a_rectangleMinY = self->top;
+	double a_rectangleMaxX = self->right;
+	double a_rectangleMaxY = self->bottom;
 	double minX = (double)a_p1x;
 	double minY;
 	double maxX = (double)a_p2x;
@@ -131,7 +131,7 @@ void psy_ui_rectangle_union(psy_ui_Rectangle* self,
 	self->bottom = self->bottom > other->bottom ? self->bottom : other->bottom;
 }
 
-void psy_ui_rectangle_expand(psy_ui_Rectangle* self, intptr_t top, intptr_t right, intptr_t bottom, intptr_t left)
+void psy_ui_rectangle_expand(psy_ui_Rectangle* self, double top, double right, double bottom, double left)
 {
 	self->top -= top;
 	self->right += right;
@@ -139,7 +139,7 @@ void psy_ui_rectangle_expand(psy_ui_Rectangle* self, intptr_t top, intptr_t righ
 	self->left -= left;	
 }
 
-void psy_ui_rectangle_move(psy_ui_Rectangle* self, intptr_t dx, intptr_t dy)
+void psy_ui_rectangle_move(psy_ui_Rectangle* self, double dx, double dy)
 {
 	self->top += dy;
 	self->right += dx;
