+++ trunk/cpsycle/ui/src/uilabel.h
@@ -22,7 +22,7 @@ extern "C" {
 
 typedef struct {
     psy_ui_Component component;
-    intptr_t charnumber;
+    double charnumber;
     double linespacing;
     psy_ui_Alignment textalignment;
     char* text;
@@ -35,7 +35,7 @@ void psy_ui_label_init_text(psy_ui_Label*, psy_ui_Component* parent,
     const char* text);
 void psy_ui_label_settext(psy_ui_Label*, const char* text);
 void psy_ui_label_text(psy_ui_Label*, char* text);
-void psy_ui_label_setcharnumber(psy_ui_Label*, intptr_t number);
+void psy_ui_label_setcharnumber(psy_ui_Label*, double number);
 void psy_ui_label_setlinespacing(psy_ui_Label*, double spacing);
 // void psy_ui_label_setstyle(psy_ui_Label*, int style);
 void psy_ui_label_settextalignment(psy_ui_Label*, psy_ui_Alignment alignment);
